docsh v0.7.1 docsh

Welcome to docsh, a documentation engine for the Erlang shell. This project provides a number of helpers for accessing module and function doc comments, function specs and exported types.

If you want your project to have embedded documentation, use rebar3_docsh plugin. If you want to browse some docs first, read on.

To access this documentation you've quite likely typed:

  h(docsh).

h/{1,2,3} is the most generic invocation of docsh. h/1 prints documentation for a given module (as shown above) or function:

  h(fun lists:keyfind/3).

h/{2,3} are limited to functions. h/2 displays information about all functions of the given name from a particular module, while h/3 also takes the expected function arity:

  h(proplists, get_value).
  h(proplists, get_value, 3).

s/{1,2,3} is the helper to use if you're only interested in function specs, not their full documentation:

  s(fun proplists:get_value/3).
  s(proplists, get_value).
  s(proplists, get_value, 3).

t/{2,3} is the helper for printing types exported from modules:

  t(gen_tcp, connect_option).
  t(gen_udp, socket).

If you want to use the helpers in erl you'll be set up by docsh's install script. It will install a user_default module enabling the shell extensions. See https://github.com/erszcz/docsh/README.md if unsure about the installation steps or what user_default is.

If you want to use docsh shell extensions in rebar3 shell, please refer to script/docsh_rebar3_shell.escript.

If you want to call docsh as a library, please refer to docsh_erl.

Types

external()

external() :: any().

Functions

activated/1

activated(path | user_default) -> ok.

load_shell_extensions/0

load_shell_extensions() -> ok.

main/1

main([string()]) -> ok.

version/0

version() -> string().