docsh v0.7.1 docsh_erl

Documentation lookup functions exported in the Erlang shell.

Types

name()

name() :: docsh_internal:name().

Functions

h/1

h(fun() | module()) -> ok.

When invoked with a module as an argument like h(lists), then look up the module documentation. When invoked with a fun as the argument like h(fun lists:keyfind/3), then equivalent to h(M, F, A) with the the fun's deduced {M :: module(), F :: name(), A :: arity()}.

h/2

h(module(), name()) -> ok.

Look up function documentation. Equivalent to h(M, F, any).

h/3

h(module(), name(), arity() | any) -> ok.

Look up function documentation. If arity is 'any', then all matching functions' docs will be printed. The complete info is function name/arity, spec and description. Use s/{1,2,3} to print just the spec.

s/1

s(fun()) -> ok.

Look up a function spec. Similar to h/1 used with a fun, but prints just the spec.

s/2

s(module(), name()) -> ok.

Look up function spec. Equivalent to s(M, F, any).

s/3

s(module(), name(), arity() | any) -> ok.

Look up function spec. If arity is 'any' print specs for all matching functions.

t/1

t(module()) -> ok.

Look up all types defined in module M.

t/2

t(module(), name()) -> ok.

Look up type definition. Equivalent to t(M, T, any).

t/3

t(module(), name(), arity() | any) -> ok.

Look up type definition. If arity is 'any' print definitions for all matching types.