docsh v0.7.1 docsh_tracer

Functions

start/0

start() -> {{tracer, pid()}, {tracer_monitor, pid()}}.

Start and monitor a tracer process.

Default dbg:dhandler/2 is nice, but when you want to use domain knowledge or discard some of the data (like too long arg lists / retvals) it might be convenient to write your own dbg handler function. However, this is error prone and since you can't trace your trace function (d'oh!) hard to debug.

The way to go is to setup a monitor on the trace handler process. When the trace function errors and kills the tracer process, you'll be notified. Still, if the monitor is setup by the shell, the 'DOWN' message might simply end up in the message queue of the shell and never be received - you won't know a crash happened.

So, it's best to setup a process which will monitor the trace handler and print any 'DOWN' messages it receives.